test: refactor to fix tests
