+++ tests/functional/test_contracts.py
@@ -1,34 +1,14 @@
 import re
-from pathlib import Path
-from typing import Optional
 
-import pytest
 from eth_utils import is_checksum_address
-from ethpm_types import ContractType
 from hexbytes import HexBytes
 
 from ape import Contract
 from ape.api import Address
-from ape.types import ContractLog
 
 SOLIDITY_CONTRACT_ADDRESS = "0xBcF7FFFD8B256Ec51a36782a52D0c34f6474D951"
 VYPER_CONTRACT_ADDRESS = "0x274b028b03A250cA03644E6c578D81f019eE1323"
 MATCH_TEST_CONTRACT = re.compile(r"<TestContract((Sol)|(Vy))")
-
-
-@pytest.fixture
-def assert_log_values(owner, chain):
-    def _assert_log_values(log: ContractLog, number: int, previous_number: Optional[int] = None):
-        assert isinstance(log.b, HexBytes)
-        expected_previous_number = number - 1 if previous_number is None else previous_number
-        assert log.prevNum == expected_previous_number, "Event param 'prevNum' has unexpected value"
-        assert log.newNum == number, "Event param 'newNum' has unexpected value"
-        assert log.dynData == "Dynamic"
-        assert log.dynIndexed == HexBytes(
-            "0x9f3d45ac20ccf04b45028b8080bb191eab93e29f7898ed43acf480dd80bba94d"
-        )
-
-    return _assert_log_values
 
 
 def test_init_at_unknown_address():
@@ -256,36 +236,6 @@ def test_call_transaction(contract_instance, owner, chain):
 
     # No mining happens because its a call
     assert init_block == chain.blocks[-1]
-
-
-def test_contract_two_events_with_same_name(owner, networks_connected_to_tester):
-    provider = networks_connected_to_tester
-    base_path = Path(__file__).parent / "data" / "contracts" / "ethereum" / "local"
-    interface_path = base_path / "Interface.json"
-    impl_path = base_path / "InterfaceImplementation.json"
-    interface_contract_type = ContractType.parse_raw(interface_path.read_text())
-    impl_contract_type = ContractType.parse_raw(impl_path.read_text())
-    event_name = "FooEvent"
-
-    # Ensure test is setup correctly in case scenario-data changed on accident
-    assert len([e for e in impl_contract_type.events if e.name == event_name]) == 2
-    assert len([e for e in interface_contract_type.events if e.name == event_name]) == 1
-
-    impl_container = provider.create_contract_container(impl_contract_type)
-    impl_instance = owner.deploy(impl_container)
-
-    with pytest.raises(AttributeError) as err:
-        _ = impl_instance.FooEvent
-
-    expected_err_prefix = f"Multiple events named '{event_name}'"
-    assert expected_err_prefix in str(err.value)
-
-    expected_sig_from_impl = "FooEvent(uint256 bar, uint256 baz)"
-    expected_sig_from_interface = "FooEvent(uint256 bar)"
-    event_from_impl_contract = impl_instance.get_event_by_signature(expected_sig_from_impl)
-    assert event_from_impl_contract.abi.signature == expected_sig_from_impl
-    event_from_interface = impl_instance.get_event_by_signature(expected_sig_from_interface)
-    assert event_from_interface.abi.signature == expected_sig_from_interface
 
 
 def test_estimating_fees(solidity_contract_instance, eth_tester_provider, owner):
